FIX STAFF_2

+++ handlers/SecondMenuHandler.py
@@ -1,5 +1,6 @@
 from telebot import types
 from .__init__ import bot,r
+from menus.FancyMenu import FancyMenu
 import pickle
 
 import sys
@@ -16,19 +17,12 @@ def unload(key):
 
 @bot.message_handler(regexp="^Сделать.*")
 def any_msg(message):
-    global user
+    # global user
     user = User(message.chat.id)
     global item
     item = Item()
     item=unload("item"+str(user.step))       #получили итем из базы по id
-
-    global keyboard
-    keyboard = types.InlineKeyboardMarkup()
-    callback_button_left = types.InlineKeyboardButton(text="<-", callback_data="to_left")
-    callback_button_ord = types.InlineKeyboardButton(text="Добавить в корзину", callback_data="create_order")
-    callback_button_right = types.InlineKeyboardButton(text="->", callback_data="to_right")
-    keyboard.add(callback_button_ord)
-    keyboard.add(callback_button_left, callback_button_right)
+    keyboard = FancyMenu()
     #bot.send_message(message.chat.id, "Выбери, что хочешь заказать: ", reply_markup=keyboard )
     bot.send_photo(message.chat.id,caption = item.description, reply_markup=keyboard, photo=item.picture)
 
@@ -40,11 +34,17 @@ def any_msg(message):
 @bot.callback_query_handler(func=lambda call: True)
 def callback_inline(call):
     # Если сообщение из чата с ботом
-    if call.message:
+    if call.message and "to" in call.data:
+        keyboard = FancyMenu()
+        if not r.get("p{}".format(call.message.chat.id)): r.set("p{}".format(call.message.chat.id),0)
+        user = User(call.message.chat.id)
+        user.step = int(r.get("p{}".format(call.message.chat.id)))
+        print(user.step)
         if call.data == "to_right":
             if user.id == call.message.chat.id:
                 if user.step >7: user.step = 0
                 user.step+=1
+                r.set("p{}".format(call.message.chat.id), user.step)
                 print("work")
                 bot.delete_message(call.message.chat.id,call.message.message_id)
                 item = unload("item" + str(user.step))
@@ -52,5 +52,13 @@ def callback_inline(call):
                                photo=item.picture)
 
                 #bot.edit_message_caption(chat_id=call.message.chat.id, message_id=call.message.message_id, caption="jiii",reply_markup=keyboard)
-
+        elif call.data == "to_left":
+            if user.id == call.message.chat.id:
+                if user.step < 2: user.step = 8
+                user.step -= 1
+                r.set("p{}".format(call.message.chat.id), user.step)
+                bot.delete_message(call.message.chat.id, call.message.message_id)
+                item = unload("item" + str(user.step))
+                bot.send_photo(call.message.chat.id, caption=item.description, reply_markup=keyboard,
+                               photo=item.picture)
